[Kaleido] Update Joshiraku 01 Part A
* (Slightly) shorten some lines for CPS
* Retime some lines

* Double down on an anachronism
This line was flagged by an outside party for being an anachronism, and not exactly fitting the time of this show. I agree. That said, I believe that subtitles are as much a product of its time as the anime themselves may be. Thus, I don't believe that taking some liberties in the name of "modernizing" is always a bad idea, like I believe to be the case here.